\name{bmuuid}
\alias{bmuuid}
\title{
Generate a probably-unique string.
}
\description{
uuid-like string generator that works whether synchronicity is available or
not.
}
\usage{bmuuid()}
\value{
The generated string
}
\author{
Luke Moryl <lsmoryl@pivotaltargeting.com>
Peter Foley <pwfoley@pivotaltargeting.com>
}

