% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%<>\%}
\alias{\%>\%}
\alias{later}
\alias{run_now}
\alias{availableCores}
\alias{\%...>\%}
\alias{\%...T>\%}
\alias{as.promise}
\alias{promise_all}
\alias{promise_map}
\alias{promise_race}
\alias{promise_reduce}
\alias{then}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{later}{\code{\link[later]{later}}, \code{\link[later]{run_now}}}

  \item{magrittr}{\code{\link[magrittr:compound]{\%<>\%}}, \code{\link[magrittr:pipe]{\%>\%}}}

  \item{parallelly}{\code{\link[parallelly]{availableCores}}}

  \item{promises}{\code{\link[promises:pipes]{\%...>\%}}, \code{\link[promises:pipes]{\%...T>\%}}, \code{\link[promises:is.promise]{as.promise}}, \code{\link[promises]{promise_all}}, \code{\link[promises]{promise_map}}, \code{\link[promises:promise_all]{promise_race}}, \code{\link[promises]{promise_reduce}}, \code{\link[promises]{then}}}
}}

